Convert lstm.
