+++ tools/adept/Makefile
@@ -5,16 +5,12 @@ LDFLAGS=-lm -ladept -fopenmp
 
 UTIL_OBJECTS=../../cpp/adbench/shared/utils.o ../../cpp/adbench/io.o
 HT_OBJECTS=AdeptHT.o run_ht.o
-LSTM_OBJECTS=AdeptLSTM.o run_lstm.o
 
 EXECUTABLES=run_hello run_gmm run_ht run_ba run_lstm
 
 all: $(EXECUTABLES)
 
 run_ht: $(HT_OBJECTS) $(UTIL_OBJECTS)
-	$(CXX) -o $@ $^ $(LDFLAGS)
-
-run_lstm: $(LSTM_OBJECTS) $(UTIL_OBJECTS)
 	$(CXX) -o $@ $^ $(LDFLAGS)
 
 run_%: run_%.cpp
@@ -24,4 +20,4 @@ run_%: run_%.cpp
 	$(CXX) $< -c -o $@ $(CFLAGS)
 
 clean:
-	rm -f $(UTIL_OBJECTS) $(LSTM_OBJECTS) $(HT_OBJECTS) $(EXECUTABLES)
+	rm -f $(UTIL_OBJECTS) $(HT_OBJECTS) $(EXECUTABLES)
