+++ tools/adol-c/Makefile
@@ -5,15 +5,11 @@ LDFLAGS=-lm -ladolc
 
 UTIL_OBJECTS=../../cpp/adbench/shared/utils.o ../../cpp/adbench/io.o
 HT_OBJECTS=AdolCHT.o run_ht.o
-LSTM_OBJECTS=AdolCLSTM.o run_lstm.o
 EXECUTABLES=run_hello run_gmm run_ba run_ht run_lstm
 
 all: $(EXECUTABLES)
 
 run_ht: $(HT_OBJECTS) $(UTIL_OBJECTS)
-	$(CXX) -o $@ $^ $(LDFLAGS)
-
-run_lstm: $(LSTM_OBJECTS) $(UTIL_OBJECTS)
 	$(CXX) -o $@ $^ $(LDFLAGS)
 
 run_%: run_%.cpp
