--- tools/manual/ManualLSTM.h
@@ -1,22 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-// https://github.com/microsoft/ADBench/blob/38cb7931303a830c3700ca36ba9520868327ac87/src/cpp/modules/manual/ManualLSTM.h
-
-#pragma once
-
-#include "adbench/shared/ITest.h"
-#include "adbench/shared/LSTMData.h"
-
-#include <vector>
-
-class ManualLSTM : public ITest<LSTMInput, LSTMOutput> {
-private:
-    std::vector<double> state;
-
-public:
-    ManualLSTM(LSTMInput&);
-
-    virtual void calculate_objective() override;
-    virtual void calculate_jacobian() override;
-};
