+++ tools/manual/lstm_d_structures.hpp
